2024/JustChess/requirements.txt
flask
python-chess
python-chess[uci,gaviota]
werkzeug
